# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TelegramLinkCodeResponse": {
        "properties": {
          "bot_username": {
            "title": "Bot Username",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "code",
          "expires_at",
          "bot_username"
        ],
        "title": "TelegramLinkCodeResponse",
        "type": "object"
      },
      "TelegramSettingsResponse": {
        "properties": {
          "is_connected": {
            "title": "Is Connected",
            "type": "boolean"
          },
          "notifications_enabled": {
            "title": "Notifications Enabled",
            "type": "boolean"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "is_connected",
          "username",
          "notifications_enabled",
          "timezone"
        ],
        "title": "TelegramSettingsResponse",
        "type": "object"
      },
      "TelegramSettingsUpdate": {
        "properties": {
          "notifications_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifications Enabled"
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "TelegramSettingsUpdate",
        "type": "object"
      },
      "TelegramTestMessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "TelegramTestMessageResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Notes API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings/telegram": {
      "get": {
        "operationId": "get_telegram_settings_api_settings_telegram_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Telegram Settings",
        "tags": [
          "telegram-settings"
        ]
      },
      "patch": {
        "operationId": "update_telegram_settings_api_settings_telegram_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Telegram Settings",
        "tags": [
          "telegram-settings"
        ]
      }
    },
    "/api/settings/telegram/link": {
      "delete": {
        "operationId": "disconnect_telegram_api_settings_telegram_link_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Disconnect Telegram",
        "tags": [
          "telegram-settings"
        ]
      }
    },
    "/api/settings/telegram/link-code": {
      "post": {
        "operationId": "create_telegram_link_code_endpoint_api_settings_telegram_link_code_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLinkCodeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Telegram Link Code Endpoint",
        "tags": [
          "telegram-settings"
        ]
      }
    },
    "/api/settings/telegram/test": {
      "post": {
        "operationId": "send_test_notification_api_settings_telegram_test_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramTestMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send Test Notification",
        "tags": [
          "telegram-settings"
        ]
      }
    },
...
  }
}